ARM: tegra: import latest Jetson TK1 pinmux
syseng has revamped the Jetson TK1 pinmux spreadsheet, basing the content
completely on correct configuration for the board/schematic, rather than
the previous version which was based on the bare minimum changes relative
to another reference board.

NOTE: THIS COMMIT IMPORTS A PRELIMINARY VERSION OF THE SPREADSHEET; IT
SHOULD NOT BE RELIED UPON OR UPSTREAMED UNTIL THE FINAL SPREADSHEET IS
AVAILABLE.

The new spreadsheet sets TRISTATE for any input-only pins. This only works
correctly if the global CLAMP bit is not set, so don't set it. Apparently
syseng have changed their mind since the previous advice that this needed
to be set:-/ In theory, we should adjust the Venice2 port to also not set
CLAMP, but I don't have time to validate that change at present.

Signed-off-by: Stephen Warren <swarren@nvidia.com>