ARM: tegra: Restrict usable RAM to 32-bit on 64-bit SoCs
Most peripherals on Tegra can do DMA only to the lower 32-bit address
space, even on 64-bit SoCs. This limitation is typically overcome by the
use of an IOMMU. Since the IOMMU is not entirely trivial to set up and
serves no other purpose (I/O protection, ...) in U-Boot, restrict 64-bit
Tegra SoCs to the lower 32-bit address space for RAM. This ensures that
the physical addresses of buffers that are programmed into the various
DMA engines are valid and don't alias to lower addresses.

Cc: Tom Warren <twarren@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>